[feat] 게임 결과 클래스 분리
